Parallel computation for the cue color data to get allRawdata.csv

+++ getProcessedResults.py
@@ -109,13 +109,13 @@ def process_filtered_data_parallel(
     # Combine all results
     allData = pd.concat(results, axis=0, ignore_index=True)
     # Ensure data types are consistent
-    allData['sub'] = allData['sub'].astype(int)
-    allData['proba'] = allData['proba'].astype(float)
-    allData['trial'] = allData['trial'].astype(int)
+    allData["sub"] = allData["sub"].astype(int)
+    allData["proba"] = allData["proba"].astype(float)
+    allData["trial"] = allData["trial"].astype(int)
 
-    events['sub'] = events['sub'].astype(int)
-    events['proba'] = events['proba'].astype(float)
-    events['trial'] = events['trial'].astype(int)
+    events["sub"] = events["sub"].astype(int)
+    events["proba"] = events["proba"].astype(float)
+    events["trial"] = events["trial"].astype(int)
 
     finalData = allData.merge(events, on=["sub", "proba", "trial"])
